Moved replication tests to a new BlockManagerReplicationSuite.
